requirements.txt

flask==3.0.3
pyyaml==6.0.2
capstone==5.0.3
scikit-learn==1.5.2
numpy==2.1.1
joblib==1.4.2
werkzeug==3.0.4
reportlab==4.2.5  # For PDF export
pandas==2.2.3  # For CSV export
